pyproject.toml
[tool.poetry]
name = "documentation-helper-llamaindex"
version = "0.1.0"
description = ""
authors = ["Antonio Henrique Machado <antoniohenriquemachado08@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
llama-index = "^0.10.59"
openai = "^1.38.0"
python-dotenv = "^1.0.1"
pinecone-client = "^5.0.1"
llama-index-vector-stores-pinecone = "^0.1.9"
unstructured = "^0.15.0"
streamlit = "^1.37.1"
llama-index-llms-groq = "^0.1.4"
llama-index-embeddings-huggingface = "^0.2.2"
transformers = "^4.44.0"


[tool.poetry.group.dev.dependencies]
ignr = "^2.2"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
